verbs: Add support for CQ moderation
This patch introduces ibv_modify_cq verb to enable CQ moderation,
it can be extended in the future for other options by using the
attr_mask field.
By using ibv_modify_cq verb a user can moderate completion events
by specifying the number of completions that cause an event
and the timeout in micro seconds to cause the event even if
the number of completions was not reached.
To disable moderation feature, a user will reset
cq_count and cq_period to 0.

Signed-off-by: Yonatan Cohen <yonatanc@mellanox.com>
Reviewed-by: Yishai Hadas <yishaih@mellanox.com>