verbs: Add flow steering egress support
Sometimes we want to carry actions when packets are sent out to the
port. In that case, we need an EGRESS based flow steering rule.
The semantics of this new EGRESS flag is port-wide, meaning - any
packet that is sent to the port is matched, regardless of the QP that
this packet originated from. The flow steering QP parameter is used
only in order to represent the device and is ignored in the matching
process.

Signed-off-by: Matan Barak <matanb@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>