debian: Let libibverbs1 16 break ibverbs-providers 15
ibverbs-providers uses the private symbols of libibverbs1. Thus
libibverbs1 will not work with an old version version of
ibverbs-providers if the private symbol version is bumped.

Bump the ibverbs-providers version that libibverbs1 breaks which was
forgotten on the symbol version bump. See also commit
611b7cc3b8e0d79aa986f042f486fccc95d73679. Append a tilde to support
the lower version numbers (e.g. 16~201711132208) in the PPA builds.

Signed-off-by: Benjamin Drung <benjamin.drung@profitbricks.com>
Signed-off-by: Leon Romanovsky <leon@kernel.org>