libhns: Make the provider sparse clean
Assuming the provider works as-is, and runs on a little endian ARM,
revise the annotations to indicate that the HW uses little endian
data in the various DMA buffers, and flow the necessary swaps throughout.

The htole/letoh swaps are no-op for this platform, which makes the only
substantive change the handling of imm_data which is now mandatory
swapped.

Signed-off-by: Jason Gunthorpe <jgunthorpe@ziepe.ca>
Signed-off-by: Lijun Ou <oulijun@huawei.com>
Signed-off-by: Leon Romanovsky <leon@kernel.org>