verbs: Introduce ibv_thread_domain and its related verbs
The ibv_thread_domain is a user space object that can be used by an
application to group verbs objects together for the purpose of
optimizing the behavior of the driver in a multi-threaded environment.

The first use of this will be for mlx5 to group UAR registers in a way
that avoids needing locking on the post send path.

This patch introduces this object and its related verbs, more details
about the usage are described as part of the man page of this commit.

Signed-off-by: Yishai Hadas <yishaih@mellanox.com>