mlx5: Add support for ibv_parent_domain
This patch adds support for ibv_alloc_parent domain verb in the mlx5
driver.

The driver uses internally the mlx5_pd structure which as its
prefix holds the core ibv_pd so that the return ibv_pd can be used
with any verb that gets a protection domain.

A parent domain must include a valid protection domain pointer and
optionally can hold a thread domain.

When used as a parent domain the core ibv_pd completely duplicates the
protection domain core fields, so the two pointers are totally
interchangeable in all core APIs, minimizing driver changes.

Signed-off-by: Yishai Hadas <yishaih@mellanox.com>