mlx5: Add support for the ibv_td object and its related verbs
This patch introduces the initial implementation of the
ibv_alloc/dealloc_td verbs.

Upon TD creation a dedicated BF (blue-flame) register should be attached
to its object and upon destruction the BF will be detached.

The motivation behind this, is to enable an application to create QP(s)
with a given TD (by using the ibv_parent_domain) and then benefit from
dropping the lock around the internal BF register upon post send.

Downstream patches in this series add the above functionality.

Signed-off-by: Yishai Hadas <yishaih@mellanox.com>