verbs: Remove container_of from the public verbs.h
This is only used in one place in the header, go ahead and open code it.
Not only does the definition pollute the global namespace, but our
inferior implementation does not do the same type checking as the
ccan version that we really want to be using, so this was causing bugs.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>