verbs: Add enum attribute type to the ioctl infrastructure
Methods sometimes need to get one attribute out of a group of
pre-defined attributes. This is an enum-like behavior. Since
this is a common requirement, we add a new ENUM attribute to the
generic ioctl layer in the kernel. This attribute is embedded in
methods, like any other attributes we currently have.

We align the user-space code with the respected changes in the kernel.

Signed-off-by: Matan Barak <matanb@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>