mlx5: Support ibv_modify_qp_rate_limit
Implement the verb ibv_modify_qp_rate_limit to support configuration of
burst control information.

The command is done through the user verbs with some driver specific
data as part of the UHW.

In case the device doesn't support burst info or rate_limit is 0
an error will be returned.

Signed-off-by: Bodong Wang <bodong@mellanox.com>
Signed-off-by: Alex Rosenbaum <alexr@mellanox.com>
Reviewed-by: Yishai Hadas <yishaih@mellanox.com>