redhat/spec: Don't include ifup/ifdown-ib
When the original Red Hat spec file was created, many things that we
used to do in the old rdma package were brought over.  One of the things
brought over were the ifup-ib and ifdown-ib network scripts.  However,
those scripts are meaningless without the SysV-initscripts package which
includes all of the sysv network startup infrastructure.  The new
rdma-core was intended to be a systemd only service, and so we skipped
the Requirements: initscripts that we should have had in order to
support these scripts.  So, remove the scripts as they are old baggage.

Fixes: dac219f1da8b ("redhat/spec: build split rpm packages")
Signed-off-by: Doug Ledford <dledford@redhat.com>