verbs: Provide infrastructure to remove kernel headers copies
The remaining kernel headers are all inlined into a struct. To avoid
major code changes we want to retain this arrangement, but still use the
kernel headers directly.

This solution is to mangle the kernel headers by script into an anonymous
and tagless struct that can be used in place of the copied data.

With the support of a few macros the whole thing becomes fairly
straightforward.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>