bnxt: Move bnxt to exclusively use the kernel uapi headers
Using the driver support macros we can now directly create the
required driver structs automatically.

There is one exciting thing here, bnxt_re has a problem with alignment
of the alloc_pd driver_data, due to a bug in the core code. bnxt worked
around it with some very ugly pointer math, but we have to solve it
properly as the common macros detect the problem.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>