mlx5: Adding support for device minimum inline header requirement
This change exposes mlx5 device minimum inline header requirements
to the mlx5 context properties and considers the inline header
requirements when preparing the eth segment of a raw packet
QP type send wqe.

Since some of mlx5 devices don't require copying the L2 header
of a raw packet into the eth segment of the wqe, we want to avoid
this copy. Specifically, when the send is performed from a device
memory buffer (DM), the address in the sg is actually an offset
which we can't use for the copy as well as we want to avoid an IO read
operation which will increase the send latency.

Signed-off-by: Ariel Levkovich <lariel@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>