umad: Do not check for umad sysfs files in umad_init
Now that we don't load the umad module if the HW doesn't use it (eg
for roce only hardware) umad_init is failing to read the ABI version
from the kernel.

Applications still want to use some libibumad services that are not
related to the char device, so move the version check to umad_open_port
instead.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>