Do not install mlx5 kernel headers if mlx5 is not built
We don't build some providers if DMA is not supported, but we still
installed the renamed kernel header in this case. This causes package
builds to fail because the sed script in debian/rules will remove the mlx5
stuff indiscriminately if DMA is not supported.

Simple fix is to be consistent and not include any mlx5 stuff if mlx5 is
not built.

Cc: stable@linux-rdma.org
Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>