Fix kernel-headers/update to work with 'git rebase -i' exec
The expected workflow for this script is to be used as part of interactive
rebase where the 'Update kernel headers' commit is dropped and replaced
with something like

  exec kernel-headers/update ../linux e818e255a58d64e86c8c93e3aa52498b1a3d1760

However, some versions of git invoke git base with GIT_DIR set in the
environment which prevents the script from reading data from the kerenl
git.

Clear harmful git environment variables before changing git repos.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>