verbs: Make fill_attr* able to handle invalid input
There are several cases where input to fill_attr comes from the untrusted
libibverbs user, and may not be marshalable to the kernel. We already have
a case like this when passing the flow counters array, but there are a few
others too.

Provide a general way for the fill_attr functions to signal un-marshaled
data and cause execute to automatically fail with EINVAL.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>