verbs: Convert write path callers to use DECLARE_LEGACY_CORE_BUFS
This changes all the write cmd wrappers that do not use UDATA to
directly use execute write and the new methods.

This removes a lot of duplicated code and makes this code closer to
something that can be moved into the ioctl fallback pattern.

As part of this patch the ibv_cmd_dealloc_mw was cleaned-up to not get
cmd and its size from its callers as there is no real usage for it in
both user space and kernel.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>