mlx5: CQE size control
When user calls ibv_create_cq the CQE size defaults to 64 bytes.
If the environment variable MLX5_CQE_SIZE is set to 64 or 128, then
CQE size will be changed accordingly.

This patch adds an option for mlx5dv_create_cq to control the CQE size
per CQ in a thread safe manner.

Signed-off-by: Yonatan Cohen <yonatanc@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>