pyverbs: Introducing pyverbs and its context class
Pyverbs is a cython package that provides Python API over IB verbs.
By that it allows both fast development of tests and an easy
introduction to RDMA for new users.

Pyverbs' development flow follows the IB verbs model. This means that
users are supposed to get ibv_context with context class, which will
provide all standard functions, e.g. query_device.

For more information, see examples section of pyverbs.

Enums need to be added as both a pxd file (for internal cython usage)
and pyx file (for Python users). To avoid code duplication, the pyx
file is a symbolic link to the pxd file.

Signed-off-by: Noa Osherovich <noaos@mellanox.com>
Signed-off-by: Leon Romanovsky <leonro@mellanox.com>