ibacm: Flush cache in provider when local address is removed
Some use cases establish IB communication within an IB port or between
local ports. That is, both the source and destination addresses could
be local to the node. When an IP address is removed,
acmp_remove_addr() destroys the address, but does not anticipate that
the address could be a local destination address as well. Hence, in
addition to deleting the source address, we iterate over all devices,
their ports, and endpoints, to see if the address we are removing also
is a local destination address. If so, we delete it from the cache.

Without this fix, we have:

 # ip addr show|grep inet' '|grep ib
    inet 192.168.200.202/24 scope global ib0
    inet 192.168.200.203/24 scope global ib1
 # ib_acme -s 192.168.200.203 -d 192.168.200.202 | egrep 192\|id:
Destination: 192.168.200.202
Source: 192.168.200.203
  dgid: fe80::21:2800:1ce:ac17
  sgid: fe80::21:2800:1ce:ac18
  dlid: 245
  slid: 246

Now, move the ib0 address to ib1:

 # ip addr del 192.168.200.202/24 dev ib0
 # ip addr add 192.168.200.202/24 dev ib1
 # ib_acme -s 192.168.200.203 -d 192.168.200.202 | egrep 192\|id:
    inet 192.168.200.203/24 scope global ib1
    inet 192.168.200.202/24 scope global secondary ib1

Check the Path Record again:

 # ib_acme -s 192.168.200.203 -d 192.168.200.202 | egrep 192\|id:
Destination: 192.168.200.202
Source: 192.168.200.203
  dgid: fe80::21:2800:1ce:ac17
  sgid: fe80::21:2800:1ce:ac18
  dlid: 245
  slid: 246

The destination addresses (dgid and dlid) are wrong, still associated
with ib0.

Signed-off-by: Håkon Bugge <haakon.bugge@oracle.com>