verbs: Replace IBV_INIT_CMD() with execute_cmd_write_resp
Replace the entire
   IBV_INIT_CMD(); write(); return errno;
Sequence with a single function call.

This signature has enough information for the execute call to compute the
location of all 5 components of a write command.

This requires adding a the _v3 version of modify_srq to the command
macros. This can be done without a lot of complexity by using core_payload
instead of the struct member aliases.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>