debian: Specify Build-Depends-Package in symbols
The symbols file does not contain a Build-Depends-Package
meta-information field.

This field specifies the name of the -dev package associated to the
library and is used by dpkg-shlibdeps(1) to make sure that the
dependency generated is at least as strict as the corresponding build
dependency.

This is useful as allows packages to not hardcode this information
multiple times.

Refer to the deb-symbols(5) manual page and the dpkg-shlibdeps(1) manual
page for details.

Signed-off-by: Benjamin Drung <benjamin.drung@cloud.ionos.com>
Signed-off-by: Leon Romanovsky <leon@kernel.org>